resolve render location using composer autoload maps (#5)
* render the base types into this repo

* csfix

* whoops

* aliases were never implemented

* mount parent directory in docker

+++ tests/bootstrap.php
@@ -1,6 +1,7 @@
 <?php declare(strict_types=1);
 
 use sndsgd\Classname;
+use sndsgd\schema\CodePathResolver;
 use sndsgd\schema\DefinedRules;
 use sndsgd\schema\DefinedTypes;
 use sndsgd\schema\RuleLocator;
@@ -34,7 +35,10 @@ function createTestTypes(string $yaml): string
     $schemaDir = BUILD_DIR . "/$hash";
     $schemaPath = $schemaDir . "/$hash.type.yaml";
 
-    mkdir($schemaDir);
+    if (!file_exists($schemaDir) && !mkdir($schemaDir, 0777, true)) {
+        throw new Exception("failed to create directory '$schemaDir'");
+    }
+
     file_put_contents($schemaPath, $yaml);
 
     $searchPaths = [$schemaDir];
@@ -69,7 +73,8 @@ function createTestTypes(string $yaml): string
         $excludePaths,
     );
 
-    $definedTypes->renderClasses(BUILD_DIR);
+    $pathResolver = new CodePathResolver(dirname(BUILD_DIR), [], BUILD_DIR);
+    $definedTypes->renderClasses($pathResolver);
 
     unlink($schemaPath);
     rmdir($schemaDir);
